The pseudohessian iteration is most likely working. It is performing slightly better than in sfbox. But more fineturning is possible, I guess. The flat hessian_width is not yet working. this is the next task.

+++ Namics/newttool.h
@@ -5,20 +5,20 @@
 #include <float.h>
 
 double norm2(double*,int);
-void decompos(double*, int, int, int&);
-void decompos(double*, int, int&);
-int signdeterminant(double*, int, int);
-int signdeterminant(double*, int);
-void multiply(double*, double, double*, double*, int, int);
-void multiply(double*, double, double*, double*, int);
-void updateneg(double*,double*, int, int, double);
-void updateneg(double* ,double* , int, double);
-void updatpos(double*, double*, double*, int, int, double);
-void updatpos(double*, double*, double*, int, double);
-void gausa(double*, double*, double*, int, int);
-void gausa(double*, double*, double*, int);
-void gausb(double*, double*, int, int);
-void gausb(double*, double*, int);
+void decompos(float*, int, int, int&);
+void decompos(float*, int, int&);
+int signdeterminant(float*, int, int);
+int signdeterminant(float*, int);
+void multiply(double*, double, float*, double*, int, int);
+void multiply(double*, double, float*, double*, int);
+void updateneg(float*,double*, int, int, double);
+void updateneg(float* ,double* , int, double);
+void updatpos(float*, double*, double*, int, int, double);
+void updatpos(float*, double*, double*, int, double);
+void gausa(float*, double*, double*, int, int);
+void gausa(float*, double*, double*, int);
+void gausb(float*, double*, int, int);
+void gausb(float*, double*, int);
 		
 
 double linecriterion(double*, double*, double*, double*, int);
